changes due to comments

+++ src/main/java/com/kodilla/ecommercee/domain/OrderDto.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,8 +16,7 @@ public class OrderDto {
     private Long id;
     private int orderNumber;
     private String productGroup;
-    private int price;
+    private BigDecimal price;
     private List<Product> productList = new ArrayList<>();
     private User user;
-
 }
